requirements.txt
streamlit
opencv-python
numpystreamlit
opencv-python
numpy
pillow
mediapipe